+++ identity-service/src/relay/txRelay.js
@@ -1,6 +1,5 @@
 const EthereumWallet = require('ethereumjs-wallet')
 const EthereumTx = require('ethereumjs-tx')
-const axios = require('axios')
 
 const models = require('../models')
 const config = require('../config')
@@ -8,13 +7,10 @@ const { logger } = require('../logging')
 
 const { AudiusABIDecoder } = require('@audius/libs')
 
-const { primaryWeb3, secondaryWeb3, ethWeb3 } = require('../web3')
+const { primaryWeb3, secondaryWeb3 } = require('../web3')
 
 // L2 relayerWallets
 const relayerConfigs = config.get('relayerWallets')
-
-// L1 relayerWallets
-const ethRelayerConfigs = config.get('ethRelayerWallets')
 
 const ENVIRONMENT = config.get('environment')
 const MIN_GAS_PRICE = config.get('minGasPrice')
@@ -142,56 +138,6 @@ const sendTransactionInternal = async (req, web3, txProps, reqBodySHA) => {
   return txReceipt
 }
 
-// Calculates index into eth relayer addresses
-const getEthRelayerWalletIndex = (walletAddress) => {
-  let walletParsedInteger = parseInt(walletAddress, 16)
-  return walletParsedInteger % ethRelayerConfigs.length
-}
-
-// Select from the list of eth relay wallet addresses
-// Return the public key that will be used to relay this address
-const selectEthRelayerWallet = (walletAddress) => {
-  return ethRelayerConfigs[getEthRelayerWalletIndex(walletAddress)].publicKey
-}
-
-// Relay a transaction to the ethereum network
-const sendEthTransaction = async (req, txProps, reqBodySHA) => {
-  const {
-    contractAddress,
-    encodedABI,
-    senderAddress,
-    gasLimit
-  } = txProps
-
-  // Calculate relayer from  senderAddressjo
-  let relayerIndex = getEthRelayerWalletIndex(senderAddress)
-  let relayerInfo = ethRelayerConfigs[relayerIndex]
-  req.logger.info(`L1 txRelay - selected relayerIndex=${relayerIndex}, relayerPublicWallet=${relayerInfo.publicKey}`)
-  let ethGasPriceInfo = await getProdGasInfo(req.app.get('redis'), req.logger)
-
-  // Select the 'fast' gas price
-  let ethRelayGasPrice = ethGasPriceInfo.fastGweiHex
-
-  let txReceipt = await createAndSendTransaction(
-    {
-      publicKey: relayerInfo.publicKey,
-      privateKey: relayerInfo.privateKey
-    },
-    contractAddress,
-    '0x00',
-    ethWeb3,
-    req.logger,
-    gasLimit,
-    encodedABI,
-    ethRelayGasPrice
-  )
-
-  txReceipt.selectedRelayIndex = relayerIndex
-  txReceipt.selectedRelay = relayerInfo
-  req.logger.info(`L1 txRelay - success, req:${reqBodySHA}, sender:${senderAddress}`)
-  return txReceipt
-}
-
 /**
  * Randomly select a wallet with a random offset. Circularly iterate through all
  * the available wallets using mod
@@ -256,35 +202,6 @@ const fundRelayerIfEmpty = async () => {
   }
 }
 
-/**
- * Fund L1 wallets as necessary to facilitate multiple relayers
- */
-const fundEthRelayerIfEmpty = async () => {
-  const minimumBalance = ethWeb3.utils.toWei(config.get('minimumBalance').toString(), 'ether')
-
-  for (let ethWallet of ethRelayerConfigs) {
-    let ethWalletPublicKey = ethWallet.publicKey
-    logger.info(`L1 Querying balance for ethRelayerPublicKey: ${ethWalletPublicKey}`)
-    let balance = await ethWeb3.eth.getBalance(ethWalletPublicKey)
-    logger.info(`L1 balance for ethRelayerPublicKey: ${balance}, minimumBalance: ${minimumBalance}`)
-    let validBalance = parseInt(balance) >= minimumBalance
-    if (ENVIRONMENT === 'development') {
-      if (!validBalance) {
-        const account = (await ethWeb3.eth.getAccounts())[0] // local acc is unlocked and does not need private key
-        logger.info(`L1 txRelay - transferring funds [${minimumBalance}] from ${account} to wallet ${ethWalletPublicKey}`)
-        await ethWeb3.eth.sendTransaction({ from: account, to: ethWalletPublicKey, value: minimumBalance })
-        logger.info(`L1 txRelay - transferred funds [${minimumBalance}] from ${account} to wallet ${ethWalletPublicKey}`)
-      } else {
-        logger.info(`L1 txRelay - ${ethWalletPublicKey} has valid balance ${balance}, minimum:${minimumBalance}`)
-      }
-    } else {
-      // In non-development environments, ethRelay wallets must be funded prior to deployment of this service
-      // Automatic funding in L1 environment is TBD
-      logger.info(`L1 txRelay -  ${ethWalletPublicKey} below minimum balance`)
-      throw new Error(`Invalid balance for ethRelayer account ${ethWalletPublicKey}. Found ${balance}, required minimumBalance ${minimumBalance}`)
-    }
-  }
-}
 
 // Send transaction using provided web3 object
 const createAndSendTransaction = async (sender, receiverAddress, value, web3, logger, gasLimit = null, data = null, inputGasPrice = null) => {
@@ -332,47 +249,9 @@ const getRelayerFunds = async (walletPublicKey) => {
   return primaryWeb3.eth.getBalance(walletPublicKey)
 }
 
-const getEthRelayerFunds = async (walletPublicKey) => {
-  return ethWeb3.eth.getBalance(walletPublicKey)
-}
-
-// Query mainnet ethereum gas prices
-const getProdGasInfo = async (redis, logger) => {
-  const prodGasPriceKey = 'eth-gas-prod-price-info'
-  let gasInfo = await redis.get(prodGasPriceKey)
-  console.log(`Found gasInfo: ${gasInfo}`)
-  if (!gasInfo) {
-    logger.info(`Redis cache miss, querying remote`)
-    let prodGasInfo = await axios({
-      method: 'get',
-      url: 'https://ethgasstation.info/api/ethgasAPI.json'
-    })
-    let { fast, fastest, safeLow, average } = prodGasInfo.data
-    gasInfo = { fast, fastest, safeLow, average }
-    // Convert returned values into gwei to be used during relay and cache
-    gasInfo.fastGwei = (parseInt(gasInfo.fast) * Math.pow(10, 9))
-    gasInfo.fastestGwei = (parseInt(gasInfo.fastest) * Math.pow(10, 9))
-    gasInfo.averageGwei = (parseInt(gasInfo.average) * Math.pow(10, 9))
-    gasInfo.fastGweiHex = ethWeb3.utils.numberToHex(gasInfo.fastGwei)
-    gasInfo.fastestGweiHex = ethWeb3.utils.numberToHex(gasInfo.fastestGwei)
-    gasInfo.averageGweiHex = ethWeb3.utils.numberToHex(gasInfo.averageGwei)
-    gasInfo.cachedResponse = false
-    redis.set(prodGasPriceKey, JSON.stringify(gasInfo), 'EX', 30)
-  } else {
-    gasInfo = JSON.parse(gasInfo)
-    gasInfo.cachedResponse = true
-  }
-  return gasInfo
-}
-
 module.exports = {
   selectWallet,
   sendTransaction,
   getRelayerFunds,
-  fundRelayerIfEmpty,
-  fundEthRelayerIfEmpty,
-  sendEthTransaction,
-  selectEthRelayerWallet,
-  getEthRelayerFunds,
-  getProdGasInfo
+  fundRelayerIfEmpty
 }
